Updated link color; Added nav element
Changed link color to orange, updated SASS file with recent CSS changes,
Made <nav> element with some <hr>'s that I want to skew and then expand
to show a list of links or something. Maybe when you hover it will
square, and on click the top of the page will slide down to show links?
That sounds pretty cool.